sdk: Restucture the details that are provided for a frame
Now, multiple formats and their associtated type of data are described.
Add the mode the camera was in when the frame was acquired.
As a result, getDataDetails() needs to be added to the API of the Frame
in order to get information about just one of the data contained by the
frame.

Signed-off-by: Dan Nechita <dan.nechita@analog.com>