sdk/connections/network: Don't wait for inexisting thread to finish
If we construct a Camera object but we don't initialize it (this happens
when we just list all available cameras) but then we destroy it, then
the Network object never gets initialized (no thread gets created). Then
the problem occurs when the Network object gets destroyed and wait for
some internal threads to finish. These threads were never created so it
ends up waiting forever.

Signed-off-by: Dan Nechita <dan.nechita@analog.com>