sdk: Fix inline function being called in other part of code
reflect8 is declared inline in it's .c file but also is being called by
other C++ class. For release builds the body of this function gets
inlined in it's .c file but for the C++ class the compiler won't know to
inline it. So the linker will complain that there is no implementation.
(LNK2019: unresolved external symbol reflect8 referenced in function
crcFast)

Signer-off-by: Dan Nechita <dan.nechita@analog.com>