sdk: Move printing SDK version from contructor of System to examples
Clients of SDK should have more flexibility to when to print the SDK
version. So don't rely on the constructor of class 'System' to print the
version.
In case of tof-viwer where the logging is rerouted to a window
that is part of the viewer, the logging of SDK version doesn't get
display as the viewer is not ready yet.

Signed-off-by: Dan Nechita <dan.nechita@analog.com>