examples/tof-viewer: Don't wait for camera ready
1st reason to remove this:
It's not clear why this is required. The function that changes the mode is a blocking type function. Once it finishes the execution, the camera should be in the new mode.

2nd reason to remove this:
The function to wait for camera calls getFrame but at the same time there is a thread that also calls getFrame. This messes things up (especially depth compute) as neither depth compute nor geFrame are thread safe.

Signed-off-by: Dan Nechita <dan.nechita@analog.com>