examples/tof-viewer: Quick (dirty) fix to make playback work with 'raw'
The frame has an internal structure (that refers to what content exists:
depth, AB, etc. and the resolution for each of them).
The current implementation of the recording process doesn't save
anything about the frame structure. So when playback starts, we don't
know how to construct the frame.
The quick fix is to assume depth and AB has been recorded.
A full refactor is needed.

Signed-off-by: Dan Nechita <dan.nechita@analog.com>