sdk: Make sure adsd3500_read_cmd overwrites the entire variable
adsd3500_read_cmd will read 4 bytes (or 2 uint16_t) and use these 4
bytes to overwrite a variable that is given. But if the variable is
wider (e.g. int, longlong, int64_t, etc.) then the rest of the most
significant bytes will remain the same. This can be a problem if the
variable is left uninitialized (thinking it will be set with a new value
anyway) since uninitialized variables have random values and most likely
the untouched bytes will be non-zero.

Signed-off-by: Dan Nechita <dan.nechita@analog.com>