sdk: Handle carefully when setting bool variable
On Windows I found the bool variable to be of size 1 byte. However the
adsd3500_read_cmd() will read 2 bytes. This caused a write outside the
memory area allocated for the bool. In my case the variable was
allocated on stack wich triggered a runtime error when
adsd3500_read_cmd() attempted to write outside the bool variable.

Signed-off-by: Dan Nechita <dan.nechita@analog.com>