bindings/python: Fix the adsd3500 getter functions
In python we can't send an int or bool or double as a reference.
Therefore we need to make the getter to return the value we want to
read. So send a pair made from status of the call and the read value.
Client code will need to read the second item from the pair to get the
value.

Signed-off-by: Dan Nechita <dan.nechita@analog.com>