sdk: Filter the types of content comming from sensor when DC on target
Remove the "raw" and "conf" that are comming from sensor when we have
Depth Compute on target. This is because sensor will only send depth and
AB.
Also rearange the frame types tables to avoid extra memcpy. Sensor is
sending Depth+AB in a continous block of data that is being copied to
the Frame object which interally has also one block of data and
different pointers pointing to "depth", "ir", "xyz", etc.

Signed-off-by: Dan Nechita <dan.nechita@analog.com>