network: Remove serializing flag. Reduce 1 frame copy on remote
Remove serializing flag. This flag was telling the remote whether the
message is wrapped in a protobuf message or is just a series of bytes.
The first byte of the message was storing the flag value. This
simplifies the server code.
Reduce one frame copy on remote by provoding the Network object with the
location where to put the incoming frame. This applies only when
expecting frame and not for protobuf messages. This seems to increase
the FPS by a couple of frames (depedens on how performant is the remote
PC).

Signed-off-by: Dan Nechita <dan.nechita@analog.com>