sdk/network: Avoid a deadlock situation when destroying Network
The scenario where deadlock occurs:
In NetworkSensorEnumerator we call searchSensors() which calls
Network::ServerConnect() which will spawn an independent thread (that
will run in a loop call_lws_service()). After spawing the thread,
ServerConnect() fails with a timeout error (waits 3 seconds). Sometimes
it timesout sometime it works. Perhaps network is slow and 3 seconds
timeout is not enough or perhaps there is another issue but not
important here. So, ServeConnect() fails and sets variable
Server_Connected to false. In parallel call_lws_serice() is executed and
sees that Server_Connected is false and exits, terminating the thread.
Then in NetworkSensorEnumerator we destroy the Network object, making
its destructor to be called. There, in that destructor, we wait for the
thead mentioned above to end (more specific, we wait that thread to
provide a signal through a condition variable). And the wait is going to
be forever since the thread no longer exists.

Signed-off-by: Dan Nechita <dan.nechita@analog.com>