examples/viewer: Make sure camera object is destroyed after 'Closed' btn
When user presses the 'Close' button, it is expected that the Camera
object is destroyed. Otherwise, the network connection (when network is
used) with the served is still active and periodic requests for updates
from server will be sent.

To do this, simplify the code by removed some structures that were
keeping alive the Camera object.

Signed-off-by: Dan Nechita <dan.nechita@analog.com>